log build variants
